Complete Stage 1
This commit contains all the changes needed to make the tests pass for Stage 1. I've removed the Healer class, as there's no way of adding a special method that affects another Character's health yet.

I also haven't added a new class: I leave that as an exercise for anyone doing this themselves!